requirements.txt
datasets==2.17.1
fortepyan==0.4.2
hydra-core==1.3.2
midi_tokenizers @ git+https://github.com/Nospoko/midi-tokenizers
pandas==2.1.4
piano_dataset @ git+https://github.com/Nospoko/piano-dataset
psycopg2==2.9.10
python-dotenv==1.0.0
sqlalchemy>=1.4.48
streamlit==1.37.0
streamlit-pianoroll==0.7.1
torch==2.5.1
